add Sccp controller

+++ src/main/java/com/unifun/sigproxy/repository/sccp/SccpAddressConfigRepository.java
@@ -3,5 +3,5 @@ package com.unifun.sigproxy.repository.sccp;
 import com.unifun.sigproxy.models.config.sccp.SccpAddressConfig;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SccpAddressConfigRepository extends JpaRepository<SccpAddressConfig, Long> {
+public interface SccpAddressConfigRepository extends JpaRepository<SccpAddressConfig, Integer> {
 }
