+++ src/main/java/com/unifun/sigproxy/repository/sccp/SccpMtp3DestinationConfigRepository.java
@@ -3,5 +3,5 @@ package com.unifun.sigproxy.repository.sccp;
 import com.unifun.sigproxy.models.config.sccp.SccpMtp3DestinationConfig;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SccpMtp3DestinationConfigRepository extends JpaRepository<SccpMtp3DestinationConfig, Long> {
+public interface SccpMtp3DestinationConfigRepository extends JpaRepository<SccpMtp3DestinationConfig, Integer> {
 }
